requirements.txt
h5py==3.6.0
keras==2.7.0
matplotlib==3.5.3
numpy==1.22.4
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
streamlit==1.11.1
tensorflow==2.7.0
opencv-python-headless


